requirements.txt
faiss-cpu
groq
langchain_groq
PyPDF2
langchain_google_genai
langchain
langchain_community
python-dotenv
pypdf
google-cloud-aiplatform>=1.38
fastapi
uvicorn
python-multipart